move instances Monad{,Atomic}Ref Fix1T to the parent
